requirements.txt
torch
torchvision
opencv-python
numpy
matplotlib
Pillow
PyYAML
